feat(linkStatic): added a flag to link a static library
As we need a static library in mobile app, added logic to generate a static library instead of the full executable.

Need to be cleaned to remove package generation that is useless.

One need to grab the .h and the .a manually to copy it to the mobile app code.